test: Add Write_Clock_Value test
